Run the tests in a docker container.
This adds two docker files and rearranges where the tests happen:

 - Dockerfile.env defines the environment needed to run the tests. It
   contains haskell, all the -devel packages we need, and all the
   dependencies from hackage. The image this defines is to be saved as
   welder/haskell-rpm-env so that the CI tests don't need to rebuild the
   dependencies each time, and so that we effectively lock the versions
   of the test dependencies until someone explicitly updates something.

 - Dockerfile.test builds on top haskell-rpm-env, and is intended to be
   rebuilt with each test. The build and test steps happen inside this
   container.

The travis.yml file, instead of containing the test scripts, is now
mostly docker setup. The yml file also contains environment variables
needed to login to docker (including an encrypted version of
DOCKER_PASSWORD=<password>), and will push a new version of
haskell-rpm-env when changes are merged to master.

Also add a .dockerignore file to make it easier to run all of this stuff
locally.